requirements.txt
streamlit
openai
plotly.express
pillow
pandas
numpy
plotly
nltk
matplotlib
seaborn
wordcloud
CleanText
scikit-learn
io